Add a mechanism to shut down the server.
Since the server runs as a daemon, it will
continue to listen on the port for connections
even after the startup script and the browser have
exited.

The original intention was to have the server terminate
automatically in the background when the user closes the
browser. This has proven to be very difficult to implement,
however, due to the following reasons:

-Attempts to tie the server "lifetime" to the execution of
the browser in the startup script have proven unsuccessful.
While it is possible for the startup script to wait until
the system() invocation of the browser has returned, this
logic will fail if an instance of the browser is already
running. In that case, the call to system() will return
immediately after a new tab has been opened on the browser.
As a consequence, system() can't be relied on to determine
whether the browser is running or not.

-Initial research has shown that the HTML DOM doesn't provide
a reliable way of identifying a "browser close" event. Generating
a client-side termination signal would thus also be complex and likely
error-prone.

-The approach of monitoring a running browser process via Unix process tables is not portable.

Because of these problems it was determined that the best
solution (for now at least) is to have the user terminate
the server via the web interface. This places slightly more
responsibility on the user than in the "ideal" scenario,
but pressing an additional button should still be a
relatively minor inconvenience.