Buffer backend messages to be served as one page.
The HTTP standard seems to dictate that the request
header and the body be transmitted in the same "chunk".
Some browsers, however, are quite lenient in regards
to this and will treat the content transmissions following
the header as belonging to the same document. Others are
more strict. Curiously, in some cases the different
OS flavors of a browser display contrary behavior.

Because of this situation, simply printing the
messages to the socket is an unreliable approach.
The new idea is to redirect STDOUT to a variable
in the child process and print the whole crap to
the socket once the replacement/restoration is done.